There has been some mix up, when omitting proxy in issue-1. So we remove proxy again
